requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
z3-solver==4.12.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyYAML==6.0.1
openai==1.3.0
anthropic==0.7.0
PyPDF2==3.0.1
python-docx==1.1.0
uvicorn==0.24.0
streamlit==1.28.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.0
httpx==0.25.2 